updated the google Auth

+++ src/main/java/easydiner/API/service/RestaurantsService.java
@@ -4,7 +4,9 @@ import easydiner.API.config.security.CustomUserDetails;
 import easydiner.API.exception.NotFoundException;
 import easydiner.API.mapper.RestaurantsMapper;
 import easydiner.API.model.RestaurantsEntity;
+import easydiner.API.model.UsersEntity;
 import easydiner.API.repository.RestaurantsRepository;
+import easydiner.API.repository.UsersRepository;
 import easydiner.API.requests.AddRestaurantRequest;
 import easydiner.API.requests.UpdateRestaurantRequest;
 import easydiner.API.responses.AddRestaurantResponse;
@@ -16,9 +18,11 @@ import lombok.extern.log4j.Log4j2;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.core.oidc.user.DefaultOidcUser;
 import org.springframework.stereotype.Service;
 
 import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 
 @Service
@@ -28,6 +32,7 @@ public class RestaurantsService {
 
     private final RestaurantsRepository restaurantsRepository;
     private final RestaurantsMapper restaurantsMapper;
+    private final UsersRepository usersRepository;
 
 
     public AddRestaurantResponse addRestaurantResponse(AddRestaurantRequest restaurantRequest) {
@@ -81,7 +86,9 @@ public class RestaurantsService {
         return getRestaurantsResponse;
     }
     public void deleteRestaurant(int id ){
-        if (restaurantExists(id)) {
+        Collection<?> roles = getAuthenticatedRole();
+        log.info("the roles are {}", roles);
+        if (hasAdminRole(roles) && restaurantExists(id)) {
             restaurantsRepository.deleteByRestaurantId(id);
         } else {
             throw new NotFoundException("Restaurant with id: "+id +" is not found");
@@ -90,14 +97,32 @@ public class RestaurantsService {
     private boolean restaurantExists(int id) {
         return restaurantsRepository.checkForExists(id);
     }
-    private int getAuthenticatedUserId() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        return ((CustomUserDetails) authentication.getPrincipal()).getUserId();
-    }
+//    private Collection getAuthenticatedRole() {
+//        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+//        return ((CustomUserDetails) authentication.getPrincipal()).getAuthorities();
+//    }
     private Collection getAuthenticatedRole() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        return ((CustomUserDetails) authentication.getPrincipal()).getAuthorities();
+        if (authentication.isAuthenticated()) {
+            Object principal = authentication.getPrincipal();
+
+            if (principal instanceof CustomUserDetails) {
+                return ((CustomUserDetails) authentication.getPrincipal()).getAuthorities();
+            } else if (principal instanceof DefaultOidcUser) {
+                DefaultOidcUser oidcUser = (DefaultOidcUser) principal;
+                String email = oidcUser.getAttribute("email");
+                log.info("inside getAuth email = {}",email);
+
+                // Use the UserService method to get the userId
+                UsersEntity usersEntity = usersRepository.findByEmail(email);
+                return Collections.singleton(usersEntity.getRole());
+            }
+            // Handle other principal types if needed
+        }
+        // Handle unauthenticated users
+        throw new IllegalStateException("User not authenticated");
     }
+
     private boolean hasAdminRole(Collection<?> roles) {
         return roles.stream()
                 .anyMatch(role -> "ROLE_ADMIN".equals(role.toString()));
